requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
rank-bm25==0.2.2
python-dotenv==1.0.1
pydantic==2.10.4
sarvamai
